+++ packages/components/atoms/f-button/test-utils/component-objects/f-button--link.component.js
@@ -1,6 +1,6 @@
 const Page = require('@justeat/f-wdio-utils/src/page.object');
 
-module.exports = class Buttons extends Page {
+module.exports = class LinkButton extends Page {
     constructor () {
         super('atom-folder', 'f-button--link-button-component');
     }