+++ packages/components/atoms/f-button/test/component/f-button.component.spec.js
@@ -1,5 +1,6 @@
 const Button = require('../../test-utils/component-objects/f-button.component');
 const LinkButton = require('../../test-utils/component-objects/f-button--link.component');
+const IconButton = require('../../test-utils/component-objects/f-button--icon.component');
 
 let button;
 
@@ -24,5 +25,16 @@ describe('f-button component tests', () => {
 
         // Assert
         expect(button.isLinkComponentDisplayed()).toBe(true);
+    });
+
+    it('should display the f-button icon component', () => {
+        // Arrange
+        button = new IconButton();
+
+        // Act
+        button.load();
+
+        // Assert
+        expect(button.isActionComponentDisplayed()).toBe(true);
     });
 });